requirements.txt
requests==2.27.1
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.36
click==8.0.4
dataclasses==0.6
Flask==1.1.2
gunicorn==20.0.4
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
typing-extensions==4.1.1
Werkzeug==2.0.3
zipp==3.6.0
configparser==5.2.0
psycopg2-binary==2.8.6
